requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.5
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.6.4
gunicorn==19.7.1
idna==2.6
mongoengine==0.13.0
pymongo==3.5.1
pytz==2017.2
requests==2.18.4
six==1.10.0
urllib3==1.22
whitenoise==3.3.0
